# Conflicts:
#	alex_flutter/adventure_it/lib/api/user_api.dart
#	alex_flutter/adventure_it/lib/constants.dart
#	alex_flutter/adventure_it/pubspec.yaml
#	backend/main-controller/pom.xml
#	backend/microservices/user-service/src/main/java/com/adventureit/userservice/Controller/UserController.java
#	backend/microservices/user-service/src/main/java/com/adventureit/userservice/Responses/GetUserByUUIDDTO.java
#	backend/microservices/user-service/src/test/java/com/adventureit/userservice/RegisterUserJunitTests.java

+++ backend/microservices/adventure-service/src/main/java/com/adventureit/adventureservice/Service/AdventureServiceImplementation.java
@@ -10,9 +10,11 @@ import com.adventureit.adventureservice.Responses.*;
 import com.adventureit.adventureservice.Exceptions.NullFieldException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import javax.transaction.Transactional;
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 @Service()
@@ -20,6 +22,8 @@ public class AdventureServiceImplementation implements AdventureService {
 
     @Autowired
     private AdventureRepository adventureRepository;
+
+    private RestTemplate restTemplate;
 
     public AdventureServiceImplementation(AdventureRepository adventureRepository){
         this.adventureRepository = adventureRepository;
@@ -50,13 +54,18 @@ public class AdventureServiceImplementation implements AdventureService {
         if(req.getOwnerId() == null ){
             throw new NullFieldException("Create Adventure Request: Owner Id NULL");
         }
-        else if (req.getId() == null) {
-            throw new NullFieldException("Create Adventure Request: Adventure Id NULL");
-        }
+//        else if (req.getId() == null) {
+//            throw new NullFieldException("Create Adventure Request: Adventure Id NULL");
+//        }
         else if (req.getName() == null){
             throw new NullFieldException("Create Adventure Request: Adventure Name NULL");
         }
-        Adventure persistedAdventure = this.adventureRepository.save(new Adventure(req.getName(),req.getDescription(), req.getId(), req.getOwnerId(), req.getStartDate(), req.getEndDate()));
+
+        //UUID location = UUID.fromString(Objects.requireNonNull(restTemplate.getForObject("http://" + "localhost" + ":" + "9006" + "/location/create/Pretoria", String.class)));
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("EEEE, dd MMMM yyyy");
+        LocalDate sd = LocalDate.parse(req.getStartDate(),formatter);
+        LocalDate ed = LocalDate.parse(req.getEndDate(),formatter);
+        Adventure persistedAdventure = this.adventureRepository.save(new Adventure(req.getName(),req.getDescription(), UUID.randomUUID() , req.getOwnerId(), sd, ed,null));
         CreateAdventureResponse response = new CreateAdventureResponse(true);
         response.setAdventure(persistedAdventure);
         return response;
@@ -134,7 +143,7 @@ public class AdventureServiceImplementation implements AdventureService {
 
         List<GetAllAdventuresResponse> list = new ArrayList<>();
         for (Adventure a:allAdventures) {
-            list.add(new GetAllAdventuresResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription()));
+            list.add(new GetAllAdventuresResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription(),a.getLocation()));
         }
 
         return list;
@@ -157,7 +166,7 @@ public class AdventureServiceImplementation implements AdventureService {
 
         List<GetAdventuresByUserUUIDResponse> list = new ArrayList<>();
         for (Adventure a:userAdventures) {
-            list.add(new GetAdventuresByUserUUIDResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription()));
+            list.add(new GetAdventuresByUserUUIDResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription(),a.getLocation()));
         }
 
         return list;
@@ -180,7 +189,7 @@ public class AdventureServiceImplementation implements AdventureService {
 
         List<GetAdventuresByUserUUIDResponse> list = new ArrayList<>();
         for (Adventure a:userAdventures) {
-            list.add(new GetAdventuresByUserUUIDResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription()));
+            list.add(new GetAdventuresByUserUUIDResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription(),a.getLocation()));
         }
 
         return list;
@@ -203,30 +212,67 @@ public class AdventureServiceImplementation implements AdventureService {
 
         List<GetAdventuresByUserUUIDResponse> list = new ArrayList<>();
         for (Adventure a:userAdventures) {
-            list.add(new GetAdventuresByUserUUIDResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription()));
+            list.add(new GetAdventuresByUserUUIDResponse(a.getId(),a.getName(),a.getAdventureId(),a.getOwnerId(),a.getAttendees(), a.getContainers(),a.getStartDate(),a.getEndDate(),a.getDescription(),a.getLocation()));
         }
 
         return list;
     }
 
     @Transactional
-    public RemoveAdventureResponse removeAdventure(UUID id){
+    public RemoveAdventureResponse removeAdventure(UUID id, UUID userID) throws Exception {
         Adventure retrievedAdventure = adventureRepository.findAdventureByAdventureId(id);
         if(retrievedAdventure == null){
             throw new AdventureNotFoundException("Remove Adventure: Adventure not found");
         }
-        adventureRepository.deleteAdventureByAdventureId(id);
+        if(!retrievedAdventure.getAttendees().contains(userID)){
+            throw new Exception("User does not belong to Adventure");
+        }
+
+        retrievedAdventure.getAttendees().remove(userID);
+
+        if(retrievedAdventure.getAttendees().isEmpty()){
+            adventureRepository.deleteAdventureByAdventureId(id);
+        }
+
         return new RemoveAdventureResponse(true, "Adventure successfully removed");
+    }
+
+    @Override
+    public List<UUID> getAttendees(UUID id) throws Exception {
+        Adventure adventure  = adventureRepository.findAdventureByAdventureId(id);
+        if(adventure == null){
+            throw new Exception("Adventure does not exist");
+        }
+
+        return adventure.getAttendees();
+    }
+
+    @Override
+    public void setAdventureLocation(UUID adventureID, UUID locationID) {
+        Adventure adventure = adventureRepository.findAdventureByAdventureId(adventureID);
+        adventure.setLocation(locationID);
+        adventureRepository.save(adventure);
+    }
+
+    @Override
+    public void addAttendees(UUID adventureID, UUID userID) throws Exception {
+        Adventure adventure = adventureRepository.findAdventureByAdventureId(adventureID);
+        if(adventure == null){
+            throw new Exception("Adventure does not exist");
+        }
+
+        adventure.getAttendees().add(userID);
+        adventureRepository.save(adventure);
     }
 
     @Override
     public void mockPopulate(){
         final UUID mockOwnerID = UUID.fromString("1660bd85-1c13-42c0-955c-63b1eda4e90b");
-        final UUID mockAttendeeID = UUID.fromString("7a984756-16a5-422e-a377-89e1772dd71e");
+        final UUID mockAttendeeID = UUID.fromString("69e8eb21-eb63-4c83-9187-181a648bb759");
 
-        Adventure mockAdventure1 = new Adventure("Mock Adventure 1","Mock Description 1 Mock Description 1Mock Description 1 Mock Description 1 Mock Description 1 Mock Description", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 7, 5),LocalDate.of(2021, 7, 9));
-        Adventure mockAdventure2 = new Adventure("Mock Adventure 2","Mock Description 2", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 1, 3),LocalDate.of(2022, 1, 2));
-        Adventure mockAdventure3 = new Adventure("Mock Adventure 3", "Mock Description 3",UUID.randomUUID(), mockAttendeeID, LocalDate.of(2022, 1, 4),LocalDate.of(2022, 1, 22));
+        Adventure mockAdventure1 = new Adventure("Mock Adventure 1","Mock Description 1 Mock Description 1Mock Description 1 Mock Description 1 Mock Description 1 Mock Description", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 7, 5),LocalDate.of(2021, 7, 9),UUID.randomUUID());
+        Adventure mockAdventure2 = new Adventure("Mock Adventure 2","Mock Description 2", UUID.randomUUID(), mockOwnerID, LocalDate.of(2021, 1, 3),LocalDate.of(2022, 1, 2),UUID.randomUUID());
+        Adventure mockAdventure3 = new Adventure("Mock Adventure 3", "Mock Description 3",UUID.randomUUID(), mockAttendeeID, LocalDate.of(2022, 1, 4),LocalDate.of(2022, 1, 22),UUID.randomUUID());
 
         mockAdventure1.addContainer(UUID.fromString("d53a7090-45f1-4eb2-953a-2258841949f8"));
         mockAdventure1.addContainer(UUID.fromString("26356837-f076-41ec-85fa-f578df7e3717"));
